Audio: Aria: Improve aria_set_config(), warn illegal blob
It depends on architecture but size_t and uint32_t are not
necessarily bits compatible (on 64bit arch like testbench
on x86). Therefore the type for att in component data of
Aria is changed to uint32_t.

A similar check as in aria_init() is added to avoid an illegal
value in the blob to break the processing algorithm.

An info level trace is added to see in trace if the blob
is applied by the Linux topology or in run-time.

+++ src/audio/aria/aria.h
@@ -64,7 +64,7 @@ struct aria_data {
 	/* current data position in circular buffer */
 	size_t buff_pos;
 	/* Attenuation parameter */
-	size_t att;
+	uint32_t att;
 	/* Gain states */
 	int32_t gains[ARIA_MAX_GAIN_STATES];
 	/* cyclic buffer pointer data */
